+++ .circleci/config.yml
@@ -198,7 +198,7 @@ jobs:
           name: "Run tests"
           environment:
             - DEBUG: "e2e:*"
-          command: yarn mocha -r ts-node/register e2e/tests/<< parameters.toml >>/*.e2e.{js,ts}
+          command: yarn mocha e2e/tests/<< parameters.toml >>/*.e2e.js
 
 workflows:
   version: 2.1
