+++ .gitignore
@@ -14,4 +14,3 @@
 
 node_modules
 deployments/
-dist/
